+++ nuxt.config.js
@@ -35,7 +35,11 @@ export default {
   css: [],
 
   // Plugins to run before rendering page (https://go.nuxtjs.dev/config-plugins)
-  plugins: ['~plugins/dayjs.ts', '~plugins/common-filter.ts'],
+  plugins: [
+    '~plugins/dayjs.ts',
+    '~plugins/common-filter.ts',
+    '~plugins/sunitaize-html.ts',
+  ],
 
   // Auto import components (https://go.nuxtjs.dev/config-components)
   components: true,
