+++ package.json
@@ -14,10 +14,12 @@
   "dependencies": {
     "@nuxt/typescript-runtime": "^2.0.0",
     "@nuxtjs/axios": "^5.12.2",
+    "@types/sanitize-html": "^1.27.0",
     "core-js": "^3.6.5",
     "dayjs": "^1.10.3",
     "nuxt": "^2.14.5",
-    "nuxt-property-decorator": "^2.8.8"
+    "nuxt-property-decorator": "^2.8.8",
+    "sanitize-html": "^2.3.0"
   },
   "devDependencies": {
     "@nuxt/types": "^2.14.5",
