+++ plugins/sunitaize-html.ts
@@ -1,0 +1,3 @@
+import Vue from 'vue'
+import sanitizeHTML from 'sanitize-html'
+Vue.prototype.$sanitize = sanitizeHTML
