+++ tsconfig.json
@@ -25,7 +25,8 @@
     },
     "types": [
       "@types/node",
-      "@nuxt/types"
+      "@nuxt/types",
+      "@types/sanitize-html"
     ]
   },
   "files": [
